+++ examples/demo/src/visitors/VisitorListAside.tsx
@@ -1,26 +1,11 @@
 import * as React from 'react';
-import { FC, ChangeEvent } from 'react';
-import {
-    Box,
-    Card,
-    CardContent,
-    makeStyles,
-    Typography,
-    IconButton,
-    InputAdornment,
-    List,
-    ListItem,
-    ListItemText,
-    ListItemSecondaryAction,
-} from '@material-ui/core';
-import SearchIcon from '@material-ui/icons/Search';
+import { FC } from 'react';
+import { Card, CardContent, makeStyles } from '@material-ui/core';
 import AccessTimeIcon from '@material-ui/icons/AccessTime';
 import MonetizationOnIcon from '@material-ui/icons/MonetizationOnOutlined';
 import MailIcon from '@material-ui/icons/MailOutline';
 import LocalOfferIcon from '@material-ui/icons/LocalOfferOutlined';
-import CancelIcon from '@material-ui/icons/CancelOutlined';
-import { Form } from 'react-final-form';
-import { TextInput, useTranslate } from 'react-admin';
+import { FilterList, FilterListItem, FilterLiveSearch } from 'react-admin';
 import {
     endOfYesterday,
     startOfWeek,
@@ -42,117 +27,38 @@ const useStyles = makeStyles(theme => ({
             display: 'none',
         },
     },
-    listItem: {
-        paddingLeft: '2em',
-    },
-    listItemText: {
-        margin: 0,
-    },
 }));
 
 const Aside: FC = props => {
-    const { filterValues, setFilters } = props as any;
     const classes = useStyles(props);
-    const translate = useTranslate();
-
-    const setFilter = (values: any) => {
-        setFilters({ ...filterValues, ...values });
-    };
-
-    const onSearchChange = (event: ChangeEvent<HTMLInputElement>) => {
-        setFilter({ q: event.target ? event.target.value : undefined });
-    };
-
-    const onSubmit = () => undefined;
-
-    // defining this component here allows to skip passing filterValues and setFilter as props
-    const FilterButton: FC<{ label: string; value: any }> = props => {
-        const { label, value } = props;
-        const isSelected = Object.keys(value).reduce(
-            (acc, key) => acc && value[key] == filterValues[key], // eslint-disable-line eqeqeq
-            true
-        );
-        const addFilter = () => {
-            if (isSelected) {
-                // remove the filter
-                const inverseValues = Object.keys(value).reduce(
-                    (acc, key) => {
-                        acc[key] = undefined;
-                        return acc;
-                    },
-                    {} as any
-                );
-                setFilter(inverseValues);
-            } else {
-                setFilter(value);
-            }
-        };
-        return (
-            <ListItem
-                button
-                onClick={addFilter}
-                selected={isSelected}
-                className={classes.listItem}
-            >
-                <ListItemText
-                    primary={translate(label)}
-                    className={classes.listItemText}
-                />
-                {isSelected && (
-                    <ListItemSecondaryAction>
-                        <IconButton size="small" onClick={addFilter}>
-                            <CancelIcon />
-                        </IconButton>
-                    </ListItemSecondaryAction>
-                )}
-            </ListItem>
-        );
-    };
 
     return (
         <Card className={classes.root}>
             <CardContent>
-                <Form onSubmit={onSubmit} initialValues={filterValues}>
-                    {({ handleSubmit }) => (
-                        <TextInput
-                            resettable
-                            helperText={false}
-                            source="q"
-                            label="Search"
-                            InputProps={{
-                                endAdornment: (
-                                    <InputAdornment position="end">
-                                        <SearchIcon color="disabled" />
-                                    </InputAdornment>
-                                ),
-                            }}
-                            onChange={onSearchChange}
-                        />
-                    )}
-                </Form>
+                <FilterLiveSearch />
 
-                <FilterSection
-                    icon={AccessTimeIcon}
+                <FilterList
                     label="resources.customers.filters.last_visited"
-                />
-                <List dense disablePadding>
-                    <FilterButton
+                    icon={<AccessTimeIcon />}
+                >
+                    <FilterListItem
+                        label="resources.customers.filters.today"
                         value={{
                             last_seen_gte: endOfYesterday().toISOString(),
                             last_seen_lte: undefined,
                         }}
-                        label="resources.customers.filters.today"
                     />
-                    <FilterButton
+                    <FilterListItem
+                        label="resources.customers.filters.this_week"
                         value={{
                             last_seen_gte: startOfWeek(
                                 new Date()
                             ).toISOString(),
                             last_seen_lte: undefined,
                         }}
-                        label="resources.customers.filters.this_week"
                     />
-                    <FilterButton
+                    <FilterListItem
+                        label="resources.customers.filters.last_week"
                         value={{
                             last_seen_gte: subWeeks(
                                 startOfWeek(new Date()),
@@ -162,18 +68,18 @@ const Aside: FC = props => {
                                 new Date()
                             ).toISOString(),
                         }}
-                        label="resources.customers.filters.last_week"
                     />
-                    <FilterButton
+                    <FilterListItem
+                        label="resources.customers.filters.this_month"
                         value={{
                             last_seen_gte: startOfMonth(
                                 new Date()
                             ).toISOString(),
                             last_seen_lte: undefined,
                         }}
-                        label="resources.customers.filters.this_month"
                     />
-                    <FilterButton
+                    <FilterListItem
+                        label="resources.customers.filters.last_month"
                         value={{
                             last_seen_gte: subMonths(
                                 startOfMonth(new Date()),
@@ -183,9 +89,9 @@ const Aside: FC = props => {
                                 new Date()
                             ).toISOString(),
                         }}
-                        label="resources.customers.filters.last_month"
                     />
-                    <FilterButton
+                    <FilterListItem
+                        label="resources.customers.filters.earlier"
                         value={{
                             last_seen_gte: undefined,
                             last_seen_lte: subMonths(
@@ -193,70 +99,51 @@ const Aside: FC = props => {
                                 1
                             ).toISOString(),
                         }}
-                        label="resources.customers.filters.earlier"
                     />
-                </List>
+                </FilterList>
 
-                <FilterSection
-                    icon={MonetizationOnIcon}
+                <FilterList
                     label="resources.customers.filters.has_ordered"
-                />
-                <List dense disablePadding>
-                    <FilterButton
+                    icon={<MonetizationOnIcon />}
+                >
+                    <FilterListItem
+                        label="ra.boolean.true"
                         value={{ has_ordered: true }}
-                        label="ra.boolean.true"
                     />
-                    <FilterButton
+                    <FilterListItem
+                        label="ra.boolean.false"
                         value={{ has_ordered: false }}
-                        label="ra.boolean.false"
                     />
-                </List>
+                </FilterList>
 
-                <FilterSection
-                    icon={MailIcon}
+                <FilterList
                     label="resources.customers.filters.has_newsletter"
-                />
-                <List dense disablePadding>
-                    <FilterButton
-                        value={{ has_newsletter: true }}
+                    icon={<MailIcon />}
+                >
+                    <FilterListItem
                         label="ra.boolean.true"
+                        value={{ has_newsletter: true }}
                     />
-                    <FilterButton
-                        value={{ has_newsletter: false }}
+                    <FilterListItem
                         label="ra.boolean.false"
+                        value={{ has_newsletter: false }}
                     />
-                </List>
+                </FilterList>
 
-                <FilterSection
-                    icon={LocalOfferIcon}
+                <FilterList
                     label="resources.customers.filters.group"
-                />
-                <List dense disablePadding>
+                    icon={<LocalOfferIcon />}
+                >
                     {segments.map(segment => (
-                        <FilterButton
-                            value={{ groups: segment.id }}
+                        <FilterListItem
                             label={segment.name}
                             key={segment.id}
+                            value={{ groups: segment.id }}
                         />
                     ))}
-                </List>
+                </FilterList>
             </CardContent>
         </Card>
-    );
-};
-
-const FilterSection: FC<{ label: string; icon: FC }> = ({
-    label,
-    icon: Icon,
-}) => {
-    const translate = useTranslate();
-    return (
-        <Box mt={2} display="flex" alignItems="center">
-            <Box mr={1}>
-                <Icon />
-            </Box>
-            <Typography variant="overline">{translate(label)}</Typography>
-        </Box>
     );
 };
