+++ packages/ra-ui-materialui/src/list/filter/index.ts
@@ -1,0 +1,5 @@
+import FilterList from './FilterList';
+import FilterListItem from './FilterListItem';
+import FilterLiveSearch from './FilterLiveSearch';
+
+export { FilterList, FilterListItem, FilterLiveSearch };
